pyproject.toml
[project]
name = "skelter"
version = "0.1.0"
description = "🦴 Skelter AI Boilerplate Generator - Plan, Scaffold, and Document your project structure in seconds."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.135.2",
    "groq>=0.30.0,<1",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-groq>=1.1.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "typer>=0.24.1",
    "uvicorn>=0.42.0",
]

[project.scripts]
skelter = "skelter.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["skelter"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
